pyproject.toml
[project]
name = "mcp-rugpull-demo"
version = "0.1.1"
description = "A demo MCP server to illustrate a Rug Pull attack in a controlled, educational way."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "CplusHua" }
]
dependencies = [
    "mcp[cli]~=1.6.0",
    "openpyxl==3.1.5",
    "pandas==2.3.0",
    "pysocks==1.7.1",
    "python-dotenv~=1.1.0",
    "requests~=2.32.3",
]

[project.scripts]
mcp-rugpull-demo = "app:mcp_rugpull"


[tool.uv]


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
